feature: изменен формат показа ошибок
